// src/@lekoarts/gatsby-theme-emilia/types.ts
export type ChildImageSharp = {
  childImageSharp: {
    fluid: {
      aspectRatio: number
      src: string
      srcSet: string
      sizes: string
      base64: string
      tracedSVG: string
      srcWebp: string
      srcSetWebp: string
      presentationHeight: number
      presentationWidth: number
      originalName: string
    }
    resize?: {
      src: string
    }
  }
}